utils: rename tests.openid_connect.pro_connect.test.py file into testing.py
Avoid any misunderstanding with tests.openid_connect.pro_connect.tests.py containing true tests
